Move mocked messages to new file
