+++ apps/meteor/tests/unit/app/apps/server/mocks/data/messages.data.js
@@ -1,3 +1,7 @@
+import { faker } from '@faker-js/faker';
+
+import { createFakeMessage } from '../../../../../../mocks/data';
+
 export const appMessageMock = {
 	id: 'appMessageMock',
 	text: 'rocket.cat',
@@ -105,3 +109,43 @@ export const appMessageInvalidRoomMock = {
 		t: 'uj',
 	},
 };
+
+const testUsername = faker.internet.userName();
+const testUserId = faker.database.mongodbObjectId();
+export const exportMessagesMock = [
+	createFakeMessage({ t: 'uj', u: { _id: testUserId, username: testUsername }, msg: testUsername }),
+	createFakeMessage({
+		msg: '',
+		file: {
+			_id: 'txt-file-id',
+			name: 'test.txt',
+			type: 'text/plain',
+			size: 29,
+			format: '',
+		},
+		attachments: [
+			{
+				type: 'file',
+				title: 'test.txt',
+				title_link: '/file-upload/txt-file-id/test.txt',
+			},
+		],
+	}),
+	createFakeMessage({
+		msg: '',
+		file: {
+			_id: 'txt-file-id',
+			name: 'test.txt',
+			type: 'text/plain',
+			size: 29,
+			format: '',
+		},
+		attachments: [
+			{
+				type: 'file',
+				title_link: '/file-upload/txt-file-id/test.txt',
+			},
+		],
+	}),
+	createFakeMessage(),
+];
